Remove cred when sign-out, or when token is expired
