+++ src/App.scss
@@ -24,9 +24,6 @@ h1, h2, h3, h4, h5 {
   color: $base;
 }
 
-.header-wrapper {
-  background-color: $background-gray;
-}
 .content-wrapper {
   display: flex;
   justify-content: center;
