+++ src/components/SubHeader/SubHeader.scss
@@ -3,6 +3,7 @@
 .subheader {
     // Top Border Line
     // border-top: solid 1px $background-gray-dark;
+    background-color: $background-gray;
 
     height: 20px;
     line-height: 18px;
