+++ src/redux/User/reducer.js
@@ -1,7 +1,8 @@
 import { UPDATE, CLEAR } from './types';
+import { __ls_remove_credentials } from 'helpers/localStorageHelpers';
 
 const INITIAL_STATE = {
-    user: null,
+    user: {user: null},
 };
 
 const reducer = (state = INITIAL_STATE, action) => {
@@ -12,11 +13,13 @@ const reducer = (state = INITIAL_STATE, action) => {
                 user: action.user,
             };
         case CLEAR:
+            __ls_remove_credentials();
             return {
                 ...state, 
-                user: null,
+                user: {user: null,}
             };
-        default: return state;
+        default: 
+            return state;
     }
 };
